+++ parley/src/main/java/nu/parley/android/data/net/ParleyRepositories.java
@@ -1,8 +1,11 @@
 package nu.parley.android.data.net;
 
 import nu.parley.android.data.repository.DeviceRepository;
+import nu.parley.android.data.repository.EventRepository;
 
 public interface ParleyRepositories {
 
     public DeviceRepository getDeviceRepository();;
+
+    public EventRepository getEventRepository();
 }
