+++ parley/src/main/java/nu/parley/android/data/repository/EventRepository.java
@@ -6,20 +6,7 @@ import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public final class EventRepository {
+public interface EventRepository {
 
-    public void fire(String event) {
-        Call<Void> eventCall = Connectivity.getRetrofit().create(EventService.class).fire(event);
-        eventCall.enqueue(new Callback<Void>() {
-            @Override
-            public void onResponse(Call<Void> call, Response<Void> response) {
-                // Ignore
-            }
-
-            @Override
-            public void onFailure(Call<Void> call, Throwable t) {
-                // Ignore
-            }
-        });
-    }
+    public void fire(String event);
 }
