+++ parley/src/main/java/nu/parley/android/view/compose/ParleyComposeListener.java
@@ -31,6 +31,6 @@ public final class ParleyComposeListener implements ComposeListener {
     }
 
     private void fireEvent(String event) {
-        new EventRepository().fire(event);
+        Parley.getInstance().getNetwork().repositories.getEventRepository().fire(event);
     }
 }
